Support "airgapped" ISOs which include copies of needed repos
The "standalone" ISO includes only the distro files, and depends
on network access to the Cobbler or remote repository server for
any repos which might be needed by the installation. If installing
on an unnetworked machine (that is, "air-gapped"), then the
installation fails when the repos are unreachable.

This commit adds a "--standalone" option to buildiso to cause it
to create a list of repos referenced by profiles in the selected
distro and copy only those repos to the ISO build directory under
a "repo_mirror" directory. When building the kickstart, the
baseurl for the repos listed need to be modified to point to
the file:// URL in the same manner as the distro URL is.