Have run() deliver a verified profile list to standalone build
The standlone build requires a distro, this adds the ability to
specify one or more child profiles under that distro to the
standalone ISO build, to allow construction of an ISO image with
only specific profiles, and thus repositories, on it. This is
helpful if you need to deliver an install-boot DVD which omits
restricted (export controlled, proprietary, etc) repos.