Change buildiso --iso description to use "file" instead of "path"
The --iso option specifies the name of a file into which the ISO
image is written. The term "path" used in the description of this
argument could be interpreted as referring to either a directory
or a file. Changing it to "file" eliminates that ambiguity.

Cherry-pick of 38f1bd0eada82d72f1cb4d8051a3194de2eac4b6
Conflicts:

	cobbler/cli.py

Merge fail due to interstitial spaces and added options; resolved.